Add support for IoT pHAT WiFi
The Red Bear IoT pHAT is a Raspberry Pi compliant HAT except for the
form factor, which fits a zero instead of the full sized models. As a
HAT, it contains EEPROM which automatically activates the firmware
overlay that it needs to utilize the MMC SDIO pins for WiFi data. This
commit adds kernel support for the WiFi half of the board only, not the
Bluetooth via UART.