Mount the ext4 partition; always try bootloader
In order to get erlang-history, the filesystem has to be mounted before
the Erlang VM starts. It looks like bootloader can't start up early
enough.

This also uncomments the bootloader start since that always works even
when bootloader.boot can't be found. This is because erlinit reverts
back to running any .boot file it finds.