Trim rootfs partitions when updating
This removes an unnecessary read/modify/write operation at the end of
firmware updates when the image doesn't fall on a 128 KB boundary.

If fwup is invoked with --enable-trim, then it will also send a TRIM
command to the underlying media. This is currently not enabled in any of
the Nerves firmware updaters, though.