Enable pstore to capture reboot causes
The pstore provides an in-memory buffer that can capture logs, kernel
oops and other information when unexpected reboots. The buffer is
recovered on the next boot where it can be inspected. This commit only
enables the drivers and allocates a 1 MB region in memory for storage.
User code is still needed to send more log information and retreive
logs.