Use nbtty to poll ttyGS0 until it exists
Since moving more drivers to modules while trying to address the USB
gadget issues, `/dev/ttyGS0` stopped being created on init. This fixes
the issue by moving the call to open the device to `nbtty` which can
poll until it has been created.

It is worthwhile to note that it appears that on the Raspberry Pi that
if you do not have anything reading from `/dev/ttyGS0`, that a Linux PC
connected over the gadget interface will lose the ability to do any USB
operations to the device if it does anything with the serial port. Many
Linux distros start ModemManager which connects to any serial port when
its detected. This was enough to hang the USB connection nearly every
time. The switch to using `nbtty` fixes this issue as well.