nerves_system_br: bump to v1.8.0 (OTP 22)
This also bumps the toolchain and Raspberry Pi kernel to match the
firmware versions provided by nerves_system_br. The device tree file
changed names in the process. According to the upstream commits, both
the new and old names are checked by the RPi bootloader and have been
for a long time.