linux: swap the cdc composite gadget for g_ether
Switching the USB gadget interface pre defined driver from USB_CDC_COMPOSITE to USB_ETH provides a more stable network connection at the expense of the virtual serial port. This adds stability to the system, while progress is made on adding support for gadget through ConfigFS. It is important to note that the UART on ttyAMA0 will contain console messages. If you were using this UART to connect to peripherials, you will need to disable the console. See: https://hexdocs.pm/nerves/advanced-configuration.html#overwriting-files-in-the-boot-partition for more information.