Pass compilation options via TARGET_GCC_FLAGS
This adds Raspberry Pi 0-specific build options for compiling ports and
NIFs. Then makes it possible for the NIF/port build options to be the
same as those used for other C/C++ code in the system (including the
Erlang runtime).

This also forces `:nerves` `v1.7.14` if using `v1.7.x`. Since
`TARGET_GCC_FLAGS` was released in a patch release (it's technically
experimental for now), it's necessary to use `v1.7.15` for this to work.